requirements.txt
reflex==0.6.4
supabase==2.10.0
python-dotenv==1.0.1